Less spaghetti in /idol
